renaming flow_stability module
