+++ src/flowstab/parallel_clustering.py
@@ -27,7 +27,7 @@ from multiprocessing import Pool, RawArray
 import numpy as np
 from scipy.sparse import csr_matrix
 
-from .FlowStability import Clustering, SparseClustering, norm_var_information
+from .flow_stability import Clustering, SparseClustering, norm_var_information
 from .sparse_stoch_mat import sparse_autocov_mat
 
 # A global dictionary storing the variables passed from the initializer.
