+++ src/flowstab/scripts/run_laplacians_transmats.py
@@ -39,7 +39,7 @@ from multiprocessing import Pool
 import numpy as np
 import pandas as pd
 
-from ..FlowStability import FlowIntegralClustering
+from ..flow_stability import FlowIntegralClustering
 from ..temporal_network import ContTempNetwork as NetClass
 
 # raise Exception
