+++ tests/test_flow_stability.py
@@ -3,83 +3,83 @@ import pytest
 def test_Partition():
     """
     """
-    from flowstab.FlowStability import Partition
+    from flowstab.flow_stability import Partition
     pass
 
 def test_BaseClustering():
     """
     """
-    from flowstab.FlowStability import BaseClustering
+    from flowstab.flow_stability import BaseClustering
     pass
 
 def test_Clustering():
     """
     """
-    from flowstab.FlowStability import Clustering
+    from flowstab.flow_stability import Clustering
     pass
 
 def test_SparseClustering():
     """
     """
-    from flowstab.FlowStability import SparseClustering
+    from flowstab.flow_stability import SparseClustering
     pass
 
 def test_FlowIntegralClustering():
     """
     """
-    from flowstab.FlowStability import FlowIntegralClustering
+    from flowstab.flow_stability import FlowIntegralClustering
     pass
 
 def test_jaccard_distance():
     """
     """
-    from flowstab.FlowStability import jaccard_distance
+    from flowstab.flow_stability import jaccard_distance
     pass
 
 def test_norm_mutual_information():
     """
     """
-    from flowstab.FlowStability import norm_mutual_information
+    from flowstab.flow_stability import norm_mutual_information
     pass
 
 def test_norm_var_information():
     """
     """
-    from flowstab.FlowStability import norm_var_information
+    from flowstab.flow_stability import norm_var_information
     pass
 
 def test_norm_var_information():
     """
     """
-    from flowstab.FlowStability import norm_var_information
+    from flowstab.flow_stability import norm_var_information
     pass
 
 def test_avg_norm_var_information():
     """
     """
-    from flowstab.FlowStability import avg_norm_var_information
+    from flowstab.flow_stability import avg_norm_var_information
     pass
 
 def test_static_clustering():
     """
     """
-    from flowstab.FlowStability import static_clustering
+    from flowstab.flow_stability import static_clustering
     pass
 
 def test_n_random_seeds():
     """
     """
-    from flowstab.FlowStability import n_random_seeds
+    from flowstab.flow_stability import n_random_seeds
     pass
 
 def test_run_multi_louvain():
     """
     """
-    from flowstab.FlowStability import run_multi_louvain
+    from flowstab.flow_stability import run_multi_louvain
     pass
 
 def test_sort_clusters():
     """
     """
-    from flowstab.FlowStability import sort_clusters
+    from flowstab.flow_stability import sort_clusters
     pass
